server/requirements.txt
djangorestframework==3.15.2
levenshtein==0.26.1
pip-chill==1.0.3
requests==2.32.3
xmltodict==0.14.2
